Change in waveform generation method

+++ music_theory/tone/oscillator.py
@@ -7,25 +7,25 @@ from .abc import OscillatorBase
 class Sinewave(OscillatorBase):
 
     def wave(self, hz: float, sec: float = 1.) -> _np.ndarray:
-        t = _np.arange(0, self.sampling_hz*sec)
+        t = _np.arange(0, int(self.sampling_hz*sec))
         return _np.sin(2 * _np.pi * hz * t / self.sampling_hz)
 
 class Suqarewave(OscillatorBase):
 
     def wave(self, hz: float, sec: float = 1.) -> _np.ndarray:
-        t = _np.arange(0, self.sampling_hz*sec)
+        t = _np.arange(0, int(self.sampling_hz*sec))
         return _sg.square(2 * _np.pi * hz * t / self.sampling_hz)
 
 class Sawtoothwave(OscillatorBase):
 
     def wave(self, hz: float, sec: float = 1.) -> _np.ndarray:
-        t = _np.arange(0, self.sampling_hz*sec)
-        return _sg.sawtooth(2 * _np.pi * hz * t / self.sampling_hz) 
+        t = _np.arange(0, int(self.sampling_hz*sec))
+        return _sg.sawtooth(2 * _np.pi * hz * t / self.sampling_hz)
 
 class Trianglewave(OscillatorBase):
 
     def wave(self, hz: float, sec: float = 1.) -> _np.ndarray:
-        t = _np.arange(0, self.sampling_hz*sec)
+        t = _np.arange(0, int(self.sampling_hz*sec))
         return _sg.sawtooth(2 * _np.pi * hz * t / self.sampling_hz, 0.5)
 
 class Noisewave(OscillatorBase):
